Created Initial Coding and Comments
Added code to complete the assignment- requested user input for file name, open the file (in read), created a list, created a loop to iterate through the file’s lines (removing the new character and whitespace) and add any words not already in the list to the list. Sort the list into alphabetical order and print.